refactor: update some import paths
