+++ src/app/entrar/services/index.ts
@@ -1,0 +1,2 @@
+export * from "./do-login";
+export * from "./do-logout";
